% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(result_daily_response)
}
\arguments{
\item{result_daily_response}{a list with three objects as produced by
\code{\link{daily_response}} function}
}
\value{
A ggplot2 object containing the heatmap display
}
\description{
Graphs a heatmpat of values stored in a matrix, such as produced
by \code{\link{daily_response}} function.
}
\examples{
data(daily_temperatures_LJ)
data(example_proxies)
Example1 <- daily_response(response = example_proxies,
env_data = daily_temperatures_LJ, method = "lm", measure = "r.squared",
fixed_width = 90, previous_year = TRUE)
plot_heatmap(Example1)

Example2 <- daily_response(response = example_proxies,
env_data = daily_temperatures_LJ, method = "brnn", measure = "adj.r.squared",
lower_limit = 50, upper_limit = 75, neurons = 1)
plot_heatmap(Example2)
}
